ml/requirements.txt
pandas
scikit-learn
cloudpickle
snowflake-connector-python
joblib
json
datetime